apps/api/requirements.txt
# FastAPI 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy>=2.0.36
asyncpg>=0.29.0
alembic==1.13.1

# Redis
redis==5.0.1
aioredis==2.0.1

# 向量数据库
qdrant-client==1.7.1

# 对象存储
boto3==1.34.14
aioboto3==12.3.0

# Gemini AI
google-genai==0.4.0

# 图像处理
pillow>=10.2.0
python-magic==0.4.27
opencv-python-headless==4.9.0.80
pillow-heif==0.15.0

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 工具
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv==1.0.0
httpx==0.26.0
numpy==1.26.3

# EXIF 解析
exifread==3.0.0

# Email Validation
email-validator>=2.0.0
